refactor: rearrange stripe dtos
