--- src/main/kotlin/ecommerce/dto/PaymentErrorResponse.kt
@@ -1,7 +1,0 @@
-package ecommerce.dto
-
-data class PaymentErrorResponse(
-    val code: String? = null,
-    val declineCode: String? = null,
-    val message: String? = null,
-)
