+++ src/main/kotlin/ecommerce/dto/stripe/PaymentResponse.kt
@@ -1,4 +1,4 @@
-package ecommerce.dto
+package ecommerce.dto.stripe
 
 class PaymentResponse(
     val id: String,
